+++ src/cwprep/__init__.py
@@ -26,7 +26,7 @@ from .config import (
     DEFAULT_CONFIG
 )
 
-__version__ = "0.2.2"
+__version__ = "0.3.0"
 __author__ = "cooper wenhua"
 __all__ = [
     "TFLBuilder",